requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
anyio==4.8.0
appdirs==1.4.4
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dateparser==1.2.1
decouple==0.0.7
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.25.2
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.1.0
narwhals==1.29.1
nest-asyncio==1.6.0
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==6.0.0
propcache==0.3.0
pycryptodome==3.21.0
pyparsing==3.2.1
PySocks==1.7.1
python-binance==1.0.28
python-dateutil==2.9.0.post0
python-telegram-bot==20.6
pytz==2025.1
regex==2024.11.6
requests==2.32.3
retrying==1.3.4
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tabulate==0.9.0
telegram-send==0.37
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wsproto==1.2.0
yarl==1.18.3
zipp==3.21.0
